Fix issue with click not making mouse_down =None
